Build script improved.

+++ build.sh
@@ -1,4 +1,3 @@
-
 #!/bin/sh
 
 # make module
@@ -27,9 +26,24 @@ cd ..
 mkdir -p append_fs/modules
 mkdir -p append_fs/home/myself
 
-cp module/kp.ko append_fs/modules/
-cp user/test  append_fs/home/myself/
-cp user/test2 append_fs/home/myself/
+if [ -e module/kp.ko ] ; then 
+    cp module/kp.ko append_fs/modules/ ;
+else
+    echo Impossible kernel module compilation failed;
+    exit 1 ;
+fi
+
+if [ -e user/test ] ; then
+    cp user/test  append_fs/home/myself/ ;
+else
+    echo "'test' compilation failed" ;
+fi
+
+if [ -e user/test2 ] ; then
+    cp user/test2 append_fs/home/myself/
+else
+    echo "'test2' compilation failed";
+fi
 
 # create install script
 echo "#!/bin/sh" > append_fs/home/myself/install_module.sh
@@ -46,11 +60,26 @@ chmod +x append_fs/home/myself/.bashrc
 cd append_fs
 find modules/ home/ | cpio -H newc -o | gzip > append_fs.gz
 cd ..
-cat TinyFS/tinyfs.gz append_fs/append_fs.gz > fs.gz
 
-# boot the machine
-qemu-system-x86_64 -kernel $KERNEL_IMAGE \
-		   -initrd fs.gz \
-		   -accel kvm \
-		   -append "console=tty consle=ttyS0" \
-		   -serial file:log.txt
+if [ $# -gt 1 ] ; then
+    initramfs=`realpath $2`;
+else
+    initramfs="TinyFS/tinyfs.gz";
+fi
+
+if [ -e $initramfs ] ; then
+    cat TinyFS/tinyfs.gz append_fs/append_fs.gz > fs.gz ;
+
+    # boot the machine
+    qemu-system-x86_64 -kernel $KERNEL_IMAGE \
+		       -initrd fs.gz \
+		       -accel kvm \
+		       -append "console=tty consle=ttyS0" \
+		       -serial file:log.txt ;
+else
+    echo I\'m not going to create an initramfs. ;
+    echo I need the initramfs file to continue. ;
+    exit 1;
+fi
+
+exit 0;
